Extract xcode-project arguments into separate file and restructure code
